requirements.txt
accelerate
bitsandbytes==0.36.0.post2
datasets
diffusers==0.25.0
ftfy==6.1.1
huggingface-hub==0.24.7
Pillow==9.4.0
python-slugify==7.0.0
torch
torchvision
transformers
safetensors
